refactor: removed commented code and redundant variables
